port fix for a segfault: https://github.com/stefan11111/kdrive/commit/0aa5ca446ec61aebcf6c10ac2f00ec61cd53cf4a
This fixes an incorrect loop condtion, as it is possible to skip over NUM_PROT when incrementing
This causes a segfault in kdrive, when km->i_prot would have to loop, instead continuing to read out of bounds.
I couldn't get tinyx to also segfault, but that doesn't mean it's impossible. It might happen on different hardware.